move new announcement modal file
move to modals folder
