+++ frontend/VolunteerOne/components/Modals/NewAnnouncementModal.js
@@ -6,7 +6,7 @@
 import React from 'react';
 import {Alert, Modal, StyleSheet, Pressable, View, Dimensions, TextInput, Image} from 'react-native';
 import { Block, Text, theme } from 'galio-framework';
-import { argonTheme } from "../constants";
+import { argonTheme } from "../../constants";
 
 const { width, height } = Dimensions.get("screen");
 
@@ -44,7 +44,7 @@ class NewAnouncementModal extends React.Component {
                 style={{alignItems: 'flex-end'}}>
                 <Image
                 style={styles.exit}
-                source={require('../assets/imgs/exit.png')} />
+                source={require('../../assets/imgs/exit.png')} />
               </Pressable>
               
               <View style={styles.modalViewInside}>
