+++ frontend/VolunteerOne/screens/Friends/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { StyleSheet, Dimensions, ScrollView } from 'react-native';
 import { Block, theme } from 'galio-framework';
 import { Button } from "../../components";
-import NewAnouncementModal from "../../components/NewAnnouncementModal"
+import NewAnouncementModal from "../../components/Modals/NewAnnouncementModal"
 import NewPostModal from '../../components/NewPostModal';
